Fixed possible race condition
Before this fix, the server reported "Done" when receiving a file after it has finished receiving. It did not wait until the file was actually written to disk, so a direct renaming of the file would fail. Now the server waits until the file has been written to disk and sends "Done" after that.