Properly close data socket after file transfer and added option to delay the 226 response
* The data socket is now closed explicitely after fetching a file

* Delay sending of code 226

An FTP client implementation has been
observed to close the data connection
as soon as it receives the 226 status
code - even though it hasn't received
all data, yet. To improve interoper
with such buggy clients, sending of
the 226 status code can now be delayed
a bit. The optional delay is controlled
by means of preprocessor define that
can be set from the cmake command
line through a cmake cache variable.

---------

Co-authored-by: Bjarne Juul Pasgaard <24828375+bjuulp@users.noreply.github.com>